requirements.txt
requests
beautifulsoup4
arxiv
google-genai
